models/classification/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.66.0
Pillow>=10.0.0
